fixed pull and cleaned up request responses

+++ server/routes.js
@@ -1,6 +1,6 @@
 import * as express from 'express';
 import * as adb from './adb.js';
-import  fileUpload from 'express-fileupload';
+import fileUpload from 'express-fileupload';
 import fs from 'fs';
 
 
@@ -13,8 +13,8 @@ router.use(fileUpload());
     [a: 10, b: 20, c: 30] => {keys: ['a', 'b', 'c'], values: [10, 20, 30]}
     Useful because express cannot send key, value array
 */
-function arrayToObject(array){
-    let keys = [];
+function arrayToObject(array) {
+    /*let keys = [];
     let values = [];
 
     for(var i in array){
@@ -22,7 +22,11 @@ function arrayToObject(array){
         values.push(array[i])
     }
 
-    return {'keys': keys, 'values': values};
+    return {'keys': keys, 'values': values};*/
+
+    let json = JSON.stringify(array);
+
+    return json;
 }
 
 
@@ -42,7 +46,7 @@ router.get('/devices', (req, res) => {
 router.post('/batterylevels', (req, res) => {
     let devices = req.body.deviceList;
     let output = adb.getBatteryLevel(devices);
-    res.send(arrayToObject(output));
+    res.send(JSON.stringify(output));
 });
 
 
@@ -53,8 +57,10 @@ router.post('/shellcmd', (req, res) => {
     let devices = req.body.deviceList; //get devices
     let cmd = req.body.cmd; //get shell command
 
+    console.log(devices);
+
     let output = adb.groupShellCmd(devices, cmd) //execute shell command and get back output
-    res.send(arrayToObject(output)); //send back output
+    res.send(JSON.stringify(output)); //send back output
 });
 
 
@@ -66,43 +72,35 @@ router.post('/adbcmd', (req, res) => {
     let cmd = req.body.cmd;
 
     let output = adb.groupAdbCmd(devices, cmd)
-    res.send(arrayToObject(output));
+    res.send(JSON.stringify(output));
 });
 
 
 /*
     Download package from front-end, install package on selected devices, delete package file (on server)
 */
-router.post('/installpackage', function (req, res) {
+router.post('/installpackage', function(req, res) {
 
     if (req.files.package == undefined || req.files.package == null) { //package wasn't uploaded properly
         res.status(415).send();
     }
-    
+
     let file = req.files.package;
     let devices = JSON.parse(req.body.deviceList);
 
 
-    console.log(file);
-
-
-    //console.log("Installing " + file.name + " on the following devices: " +  devices);
-
-    
     if (file.mimetype == 'application/vnd.android.package-archive') { //check if package is an .apk
-        file.mv(file.name, function (error, response) {//save file
+        file.mv(file.name, function(error, response) { //save file
             if (error) {
                 res.status(415).send();
-            }
-            else { //file was successfully saved
+            } else { //file was successfully saved
 
                 let output = adb.installPackage(devices, file.name); //install package
-                res.send(arrayToObject(output));
-                fs.rm(file.name, ()=>{}); //delete file (no callback needed)
+                res.send(JSON.stringify(output));
+                fs.rm(file.name, () => {}); //delete file (no callback needed)
             }
         });
-    }
-    else {
+    } else {
         res.status(415).send("Wrong file type");
     }
 });
@@ -117,11 +115,11 @@ router.post('/checkpackageinstalled', (req, res) => {
 
     let output = [];
 
-    for(var i in devices){
+    for (var i in devices) {
         output[devices[i]] = adb.checkPackageInstalled(devices[i], packageName);
     }
 
-    res.send(arrayToObject(output));
+    res.send(JSON.stringify(output));
 });
 
 
@@ -132,7 +130,7 @@ router.post('/installedpackages', (req, res) => {
     let devices = req.body.deviceList;
 
     let output = adb.getPackages(devices);
-    res.send(arrayToObject(output));
+    res.send(JSON.stringify(output));
 });
 
 
@@ -145,35 +143,34 @@ router.post('/uninstallpackage', (req, res) => {
 
     let output = adb.uninstallPackage(devices, packageName);
 
-    res.send(arrayToObject(output));
+    res.send(JSON.stringify(output));
 });
 
 
 /*
     Download file from front-end, push to selected devices (root folder), delete file (on server)
 */
-router.post('/pushfile', function (req, res) {
+router.post('/pushfile', function(req, res) {
 
     if (req.files.file == undefined || req.files.file == null) { //file wasn't uploaded properly
         res.status(400).send("No file uploaded");
     }
-    
+
     let file = req.files.file;
     let devices = JSON.parse(req.body.deviceList);
 
-    
-    file.mv(file.name, function (error, response) {//save file
+
+    file.mv(file.name, function(error, response) { //save file
         if (error) {
             res.status(400).send();
-        }
-        else { //file was successfully saved
+        } else { //file was successfully saved
 
             let output = adb.pushFiles(devices, file.name, "/sdcard/"); //push file
-            res.send(arrayToObject(output));
-            fs.rm(file.name, ()=>{}); //delete file (no callback needed)
+            res.send(JSON.stringify(output));
+            fs.rm(file.name, () => {}); //delete file (no callback needed)
         }
     });
-    
+
 });
 
 
@@ -186,40 +183,60 @@ router.post('/deletefile', (req, res) => {
 
     let output = adb.deleteFile(devices, filePath); //no feedback
 
-    res.send(arrayToObject(output));
+    res.send(JSON.stringify(output));
 });
 
 
 /*
-    Pull all files, then
-
-    For each device: 
-        rename the file to deviceName_fileName.ext
-        convert the file data to 64 bit -> save in array
-        delete the file
+    For each device:
+        attempt to pull the file
+        if successful then
+            convert the file data to 64 bit -> save in array
+            delete the file from the server
+        if unsuccessful then
+            send back and error
     Send the array 
 */
 router.post('/pullfile', (req, res) => {
     let devices = req.body.deviceList;
     let filePath = req.body.filePath;
-    let filesOutput = []; //array containing [filename : 64bit_data, ...]
 
     let splitFilePath = filePath.split("/"); //get filename from source
-    let fileName = splitFilePath[splitFilePath.length-1]
+    let fileName = splitFilePath[splitFilePath.length - 1]
 
-    let output = adb.pullFiles(devices, filePath);
+    let output = {}; //object containing {device1 : {success: true, data: 64bit_data, filename: filname}, device2 : {success: false, error: error message}, ...}
 
-    for(var i in devices){
-        let newFileName = devices[i]+"_"+fileName
-        
-        let data = fs.readFileSync(newFileName, {encoding:'base64', flag:'r'});
+    for (var i in devices) {
+        let success = adb.pullFiles(devices[i], filePath);
 
-        filesOutput[newFileName] = data
-
-        fs.rm(newFileName, ()=>{}); //delete file (no callback needed)
+        if (success) {
+            let newFileName = devices[i] + "_" + fileName; //file is renamed during pull
+            let data = fs.readFileSync(newFileName, { encoding: 'base64', flag: 'r' });
+            output[devices[i]] = { success: true, data: data, filename: newFileName };
+            fs.rm(newFileName, () => {}); //delete file (no callback needed)
+        } else {
+            output[devices[i]] = { success: false, error: "File " + filePath + " does not exist on device " + devices[i] };
+        }
     }
 
-    res.send(arrayToObject(filesOutput));
+    console.log(output);
+    console.log(JSON.stringify(output));
+
+
+    res.send(JSON.stringify(output));
+
+
+
+
+    /*
+
+    for each device:
+
+        output = {devicename: {success, data/error}, }
+
+        pull file
+    
+    */
 });
 
 
@@ -233,13 +250,13 @@ router.post('/addwifi', (req, res) => {
     let passwordType = req.body.passwordType;
     let password = req.body.password;
 
-    if(passwordType == ""){
+    if (passwordType == "") {
         passwordType = "none"
     }
 
     let output = adb.addWifiNetwork(devices, ssid, passwordType, password);
-    
-    res.send(arrayToObject(output));
+
+    res.send(JSON.stringify(output));
 });
 
 
@@ -249,7 +266,7 @@ router.post('/addwifi', (req, res) => {
 router.post('/disablewifi', (req, res) => {
     let devices = req.body.deviceList;
     let output = adb.toggleWifi(devices, false);
-    res.send(arrayToObject(output));
+    res.send(JSON.stringify(output));
 });
 
 
@@ -259,87 +276,76 @@ router.post('/disablewifi', (req, res) => {
 router.post('/enablewifi', (req, res) => {
     let devices = req.body.deviceList;
     let output = adb.toggleWifi(devices, true);
-    res.send(arrayToObject(output));
+    res.send(JSON.stringify(output));
 });
 
 
 /*
 
 */
-router.post('/forgetallwifi', (req, res) =>{
+router.post('/forgetallwifi', (req, res) => {
     let devices = req.body.deviceList;
     let output = adb.forgetWifi(devices);
-    res.send(arrayToObject(output));
+    res.send(JSON.stringify(output));
 })
 
 
 /*
 
 */
-router.post('/recordscreen', (req, res) =>{
+router.post('/recordscreen', (req, res) => {
     let devices = req.body.deviceList;
     let seconds = req.body.seconds
     let filesOutput = []; //array containing [filename : 64bit_data, ...]
     let filenames = [];
 
-    for(var i in devices){
+    for (var i in devices) {
         filenames.push(adb.recordScreen(devices[i], seconds));
     }
 
-    for(var i in filenames){
-        
-        let data = fs.readFileSync(filenames[i], {encoding:'base64', flag:'r'});
+    for (var i in filenames) {
+
+        let data = fs.readFileSync(filenames[i], { encoding: 'base64', flag: 'r' });
 
         filesOutput[filenames[i]] = data
 
-        fs.rm(filenames[i], ()=>{}); //delete file (no callback needed)
+        fs.rm(filenames[i], () => {}); //delete file (no callback needed)
     }
 
-    res.send(arrayToObject(filesOutput));
+    res.send(JSON.stringify(filesOutput));
 })
 
 
 /*
 
 */
-router.post('/screencapture', (req, res) =>{
+router.post('/screencapture', (req, res) => {
     let devices = req.body.deviceList;
     let filesOutput = []; //array containing [filename : 64bit_data, ...]
     let filenames = [];
 
-    for(var i in devices){
+    for (var i in devices) {
         filenames.push(adb.screenCap(devices[i]));
     }
 
-    for(var i in filenames){
-        
-        let data = fs.readFileSync(filenames[i], {encoding:'base64', flag:'r'});
+    for (var i in filenames) {
+
+        let data = fs.readFileSync(filenames[i], { encoding: 'base64', flag: 'r' });
 
         filesOutput[filenames[i]] = data
 
-        fs.rm(filenames[i], ()=>{}); //delete file (no callback needed)
+        fs.rm(filenames[i], () => {}); //delete file (no callback needed)
     }
 
-    res.send(arrayToObject(filesOutput));
+    res.send(JSON.stringify(filesOutput));
 })
 
 
 /*
 
 */
-router.post('/getwificonnection', (req, res) =>{
+router.post('/getwificonnection', (req, res) => {
     let devices = req.body.deviceList;
     let output = adb.checkWifiNetwork(devices);
-    res.send(arrayToObject(output));
+    res.send(JSON.stringify(output));
 })
-
-
-
-
-
-
-
-
-
-
-
